+++ tests/test_main.py
@@ -10,6 +10,7 @@
 
 import errno
 import os
+import re
 from pathlib import Path
 from typing import Optional
 from unittest.mock import create_autospec
@@ -100,6 +101,15 @@ def test_lint_fail(fake_repository, stringio):
     assert result > 0
     assert "foo.py" in stringio.getvalue()
     assert ":-(" in stringio.getvalue()
+
+
+def test_lint_fail_quiet(fake_repository, stringio):
+    """Run a failed lint."""
+    (fake_repository / "foo.py").write_text("foo")
+    result = main(["lint", "--quiet"], out=stringio)
+
+    assert result > 0
+    assert stringio.getvalue() == ""
 
 
 def test_lint_no_file_extension(fake_repository, stringio):
@@ -251,3 +261,15 @@ def test_download_custom_output_too_many(
         main(
             ["download", "-o", "foo", "0BSD", "GPL-3.0-or-later"], out=stringio
         )
+
+
+def test_supported_licenses(stringio):
+    """Invoke the supported-licenses command and check whether the result
+    contains at least one license in the expected format.
+    """
+
+    assert main(["supported-licenses"], out=stringio) == 0
+    assert re.search(
+        r"GPL-3\.0-or-later\s+GNU General Public License v3\.0 or later\s+https:\/\/spdx\.org\/licenses\/GPL-3\.0-or-later\.html\s+\n",
+        stringio.getvalue(),
+    )