+++ CHANGELOG.md
@@ -40,6 +40,8 @@ The versions follow [semantic versioning](https://semver.org).
 ### Added
 
 - More file types are recognised:
+  - Protobuf files (`.proto`)
+  - GraphQL (`.graphql`)
   - Perl plain old documentation (`.pod`)
   - Portable document format (`.pdf`)
   - Open Document Format (`.odt`, `.ods`, `.fodp` and many more)
@@ -55,6 +57,9 @@ The versions follow [semantic versioning](https://semver.org).
 
 - More file names are recognised:
   - SuperCollider (`archive.sctxar`)
+
+- `--quiet` switch to the `lint` command
+- `supported-licenses` command that lists all licenses supported by REUSE
 
 ### Changed